make asnyc first class and port tests and examples
