+++ Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "tar"
-version = "0.4.26"
-authors = ["Alex Crichton <alex@alexcrichton.com>"]
+name = "async-tar"
+version = "0.1.0"
+authors = ["dignifiedquire <me@dignifiequire.com>", "Alex Crichton <alex@alexcrichton.com>"]
 homepage = "https://github.com/alexcrichton/tar-rs"
 repository = "https://github.com/alexcrichton/tar-rs"
-documentation = "https://docs.rs/tar"
+documentation = "https://docs.rs/async-tar"
 license = "MIT/Apache-2.0"
 keywords = ["tar", "tarfile", "encoding"]
 readme = "README.md"
@@ -26,6 +26,7 @@ pin-cell = "0.1.1"
 
 [dev-dependencies]
 tempfile = "3"
+async-std = {version = "1.4.0", features = ["unstable", "attributes"] }
 
 [target."cfg(unix)".dependencies]
 xattr = { version = "0.2", optional = true }
