+++ src/entry.rs
@@ -1,15 +1,18 @@
 use std::borrow::Cow;
 use std::cmp;
-use std::fs;
-use std::fs::OpenOptions;
-use std::io::prelude::*;
-use std::io::{self, Error, ErrorKind, SeekFrom};
 use std::marker;
-use std::path::{Component, Path, PathBuf};
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use async_std::fs;
+use async_std::fs::OpenOptions;
+use async_std::io::prelude::*;
+use async_std::io::{self, Error, ErrorKind, SeekFrom};
+use async_std::path::{Component, Path, PathBuf};
+use pin_project::{pin_project, project};
 
 use filetime::{self, FileTime};
 
-use crate::archive::ArchiveInner;
 use crate::error::TarError;
 use crate::header::bytes2path;
 use crate::other;
@@ -21,14 +24,17 @@ use crate::{Archive, Header, PaxExtensions};
 /// This structure is a window into a portion of a borrowed archive which can
 /// be inspected. It acts as a file handle by implementing the Reader trait. An
 /// entry cannot be rewritten once inserted into an archive.
-pub struct Entry<'a, R: 'a + Read> {
-    fields: EntryFields<'a>,
-    _ignored: marker::PhantomData<&'a Archive<R>>,
+#[pin_project]
+pub struct Entry<R: Read + Unpin> {
+    #[pin]
+    fields: EntryFields<R>,
+    _ignored: marker::PhantomData<Archive<R>>,
 }
 
 // private implementation detail of `Entry`, but concrete (no type parameters)
 // and also all-public to be constructed from other modules.
-pub struct EntryFields<'a> {
+#[pin_project]
+pub struct EntryFields<R: Read + Unpin> {
     pub long_pathname: Option<Vec<u8>>,
     pub long_linkname: Option<Vec<u8>>,
     pub pax_extensions: Option<Vec<u8>>,
@@ -36,15 +42,19 @@ pub struct EntryFields<'a> {
     pub size: u64,
     pub header_pos: u64,
     pub file_pos: u64,
-    pub data: Vec<EntryIo<'a>>,
+    #[pin]
+    pub data: Vec<EntryIo<R>>,
     pub unpack_xattrs: bool,
     pub preserve_permissions: bool,
     pub preserve_mtime: bool,
+    #[pin]
+    pub(crate) read_state: Option<EntryIo<R>>,
 }
 
-pub enum EntryIo<'a> {
-    Pad(io::Take<io::Repeat>),
-    Data(io::Take<&'a ArchiveInner<dyn Read + 'a>>),
+#[pin_project]
+pub enum EntryIo<R: Read + Unpin> {
+    Pad(#[pin] io::Take<io::Repeat>),
+    Data(#[pin] io::Take<R>),
 }
 
 /// When unpacking items the unpacked thing is returned to allow custom
@@ -53,13 +63,13 @@ pub enum EntryIo<'a> {
 #[derive(Debug)]
 pub enum Unpacked {
     /// A file was unpacked.
-    File(std::fs::File),
+    File(fs::File),
     /// A directory, hardlink, symlink, or other node was unpacked.
     #[doc(hidden)]
     __Nonexhaustive,
 }
 
-impl<'a, R: Read> Entry<'a, R> {
+impl<R: Read + Unpin> Entry<R> {
     /// Returns the path name for this entry.
     ///
     /// This method may fail if the pathname is not valid Unicode and this is
@@ -130,8 +140,8 @@ impl<'a, R: Read> Entry<'a, R> {
     ///
     /// Also note that this function will read the entire entry if the entry
     /// itself is a list of extensions.
-    pub fn pax_extensions(&mut self) -> io::Result<Option<PaxExtensions>> {
-        self.fields.pax_extensions()
+    pub async fn pax_extensions(&mut self) -> io::Result<Option<PaxExtensions<'_>>> {
+        self.fields.pax_extensions().await
     }
 
     /// Returns access to the header of this entry in the archive.
@@ -190,8 +200,8 @@ impl<'a, R: Read> Entry<'a, R> {
     ///     file.unpack(format!("file-{}", i)).unwrap();
     /// }
     /// ```
-    pub fn unpack<P: AsRef<Path>>(&mut self, dst: P) -> io::Result<Unpacked> {
-        self.fields.unpack(None, dst.as_ref())
+    pub async fn unpack<P: AsRef<Path>>(&mut self, dst: P) -> io::Result<Unpacked> {
+        self.fields.unpack(None, dst.as_ref()).await
     }
 
     /// Extracts this file under the specified path, avoiding security issues.
@@ -218,8 +228,8 @@ impl<'a, R: Read> Entry<'a, R> {
     ///     file.unpack_in("target").unwrap();
     /// }
     /// ```
-    pub fn unpack_in<P: AsRef<Path>>(&mut self, dst: P) -> io::Result<bool> {
-        self.fields.unpack_in(dst.as_ref())
+    pub async fn unpack_in<P: AsRef<Path>>(&mut self, dst: P) -> io::Result<bool> {
+        self.fields.unpack_in(dst.as_ref()).await
     }
 
     /// Indicate whether extended file attributes (xattrs on Unix) are preserved
@@ -251,32 +261,52 @@ impl<'a, R: Read> Entry<'a, R> {
     }
 }
 
-impl<'a, R: Read> Read for Entry<'a, R> {
-    fn read(&mut self, into: &mut [u8]) -> io::Result<usize> {
-        self.fields.read(into)
+impl<R: Read + Unpin> Read for Entry<R> {
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        into: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        let mut this = self.project();
+        Pin::new(&mut *this.fields).poll_read(cx, into)
     }
 }
 
-impl<'a> EntryFields<'a> {
-    pub fn from<R: Read>(entry: Entry<R>) -> EntryFields {
+impl<R: Read + Unpin> EntryFields<R> {
+    pub fn from(entry: Entry<R>) -> Self {
         entry.fields
     }
 
-    pub fn into_entry<R: Read>(self) -> Entry<'a, R> {
+    pub fn into_entry(self) -> Entry<R> {
         Entry {
             fields: self,
             _ignored: marker::PhantomData,
         }
     }
 
-    pub fn read_all(&mut self) -> io::Result<Vec<u8>> {
+    pub(crate) fn poll_read_all(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<io::Result<Vec<u8>>> {
+        // Preallocate some data but don't let ourselves get too crazy now.
+        let cap = cmp::min(self.size, 128 * 1024);
+        let mut buf = Vec::with_capacity(cap as usize);
+
+        // Copied from futures::ReadToEnd
+        match async_std::task::ready!(poll_read_all_internal(self, cx, &mut buf)) {
+            Ok(_) => Poll::Ready(Ok(buf)),
+            Err(err) => Poll::Ready(Err(err)),
+        }
+    }
+
+    pub async fn read_all(&mut self) -> io::Result<Vec<u8>> {
         // Preallocate some data but don't let ourselves get too crazy now.
         let cap = cmp::min(self.size, 128 * 1024);
         let mut v = Vec::with_capacity(cap as usize);
-        self.read_to_end(&mut v).map(|_| v)
+        self.read_to_end(&mut v).await.map(|_| v)
     }
 
-    fn path(&self) -> io::Result<Cow<Path>> {
+    fn path(&self) -> io::Result<Cow<'_, Path>> {
         bytes2path(self.path_bytes())
     }
 
@@ -329,19 +359,19 @@ impl<'a> EntryFields<'a> {
         }
     }
 
-    fn pax_extensions(&mut self) -> io::Result<Option<PaxExtensions>> {
+    async fn pax_extensions(&mut self) -> io::Result<Option<PaxExtensions<'_>>> {
         if self.pax_extensions.is_none() {
             if !self.header.entry_type().is_pax_global_extensions()
                 && !self.header.entry_type().is_pax_local_extensions()
             {
                 return Ok(None);
             }
-            self.pax_extensions = Some(self.read_all()?);
+            self.pax_extensions = Some(self.read_all().await?);
         }
         Ok(Some(pax_extensions(self.pax_extensions.as_ref().unwrap())))
     }
 
-    fn unpack_in(&mut self, dst: &Path) -> io::Result<bool> {
+    async fn unpack_in(&mut self, dst: &Path) -> io::Result<bool> {
         // Notes regarding bsdtar 2.8.3 / libarchive 2.8.3:
         // * Leading '/'s are trimmed. For example, `///test` is treated as
         //   `test`.
@@ -393,45 +423,49 @@ impl<'a> EntryFields<'a> {
             None => return Ok(false),
         };
 
-        if parent.symlink_metadata().is_err() {
-            fs::create_dir_all(&parent).map_err(|e| {
+        if parent.symlink_metadata().await.is_err() {
+            fs::create_dir_all(&parent).await.map_err(|e| {
                 TarError::new(&format!("failed to create `{}`", parent.display()), e)
             })?;
         }
 
-        let canon_target = self.validate_inside_dst(&dst, parent)?;
+        let canon_target = self.validate_inside_dst(&dst, parent).await?;
 
         self.unpack(Some(&canon_target), &file_dst)
+            .await
             .map_err(|e| TarError::new(&format!("failed to unpack `{}`", file_dst.display()), e))?;
 
         Ok(true)
     }
 
     /// Unpack as destination directory `dst`.
-    fn unpack_dir(&mut self, dst: &Path) -> io::Result<()> {
+    async fn unpack_dir(&mut self, dst: &Path) -> io::Result<()> {
         // If the directory already exists just let it slide
-        fs::create_dir(dst).or_else(|err| {
-            if err.kind() == ErrorKind::AlreadyExists {
-                let prev = fs::metadata(dst);
-                if prev.map(|m| m.is_dir()).unwrap_or(false) {
-                    return Ok(());
+        match fs::create_dir(dst).await {
+            Ok(()) => Ok(()),
+            Err(err) => {
+                if err.kind() == ErrorKind::AlreadyExists {
+                    let prev = fs::metadata(dst).await;
+                    if prev.map(|m| m.is_dir()).unwrap_or(false) {
+                        return Ok(());
+                    }
                 }
+                Err(Error::new(
+                    err.kind(),
+                    format!("{} when creating dir {}", err, dst.display()),
+                ))
             }
-            Err(Error::new(
-                err.kind(),
-                format!("{} when creating dir {}", err, dst.display()),
-            ))
-        })
+        }
     }
 
     /// Returns access to the header of this entry in the archive.
-    fn unpack(&mut self, target_base: Option<&Path>, dst: &Path) -> io::Result<Unpacked> {
+    async fn unpack(&mut self, target_base: Option<&Path>, dst: &Path) -> io::Result<Unpacked> {
         let kind = self.header.entry_type();
 
         if kind.is_dir() {
-            self.unpack_dir(dst)?;
+            self.unpack_dir(dst).await?;
             if let Ok(mode) = self.header.mode() {
-                set_perms(dst, None, mode, self.preserve_permissions)?;
+                set_perms(dst, None, mode, self.preserve_permissions).await?;
             }
             return Ok(Unpacked::__Nonexhaustive);
         } else if kind.is_hard_link() || kind.is_symlink() {
@@ -467,12 +501,12 @@ impl<'a> EntryFields<'a> {
                     // so we need to validate at this time.
                     Some(ref p) => {
                         let link_src = p.join(src);
-                        self.validate_inside_dst(p, &link_src)?;
+                        self.validate_inside_dst(p, &link_src).await?;
                         link_src
                     }
                     None => src.into_owned(),
                 };
-                fs::hard_link(&link_src, dst).map_err(|err| {
+                fs::hard_link(&link_src, dst).await.map_err(|err| {
                     Error::new(
                         err.kind(),
                         format!(
@@ -525,9 +559,9 @@ impl<'a> EntryFields<'a> {
         // Names that have a trailing slash should be treated as a directory.
         // Only applies to old headers.
         if self.header.as_ustar().is_none() && self.path_bytes().ends_with(b"/") {
-            self.unpack_dir(dst)?;
+            self.unpack_dir(dst).await?;
             if let Ok(mode) = self.header.mode() {
-                set_perms(dst, None, mode, self.preserve_permissions)?;
+                set_perms(dst, None, mode, self.preserve_permissions).await?;
             }
             return Ok(Unpacked::__Nonexhaustive);
         }
@@ -543,74 +577,82 @@ impl<'a> EntryFields<'a> {
 
         // Ensure we write a new file rather than overwriting in-place which
         // is attackable; if an existing file is found unlink it.
-        fn open(dst: &Path) -> io::Result<std::fs::File> {
-            OpenOptions::new().write(true).create_new(true).open(dst)
+        async fn open(dst: &Path) -> io::Result<fs::File> {
+            OpenOptions::new()
+                .write(true)
+                .create_new(true)
+                .open(dst)
+                .await
         };
-        let mut f = (|| -> io::Result<std::fs::File> {
-            let mut f = open(dst).or_else(|err| {
-                if err.kind() != ErrorKind::AlreadyExists {
-                    Err(err)
-                } else {
-                    match fs::remove_file(dst) {
-                        Ok(()) => open(dst),
-                        Err(ref e) if e.kind() == io::ErrorKind::NotFound => open(dst),
-                        Err(e) => Err(e),
+        let mut f = async {
+            let mut f = match open(dst).await {
+                Ok(f) => Ok(f),
+                Err(err) => {
+                    if err.kind() != ErrorKind::AlreadyExists {
+                        Err(err)
+                    } else {
+                        match fs::remove_file(dst).await {
+                            Ok(()) => open(dst).await,
+                            Err(ref e) if e.kind() == io::ErrorKind::NotFound => open(dst).await,
+                            Err(e) => Err(e),
+                        }
                     }
                 }
-            })?;
+            }?;
             for io in self.data.drain(..) {
                 match io {
                     EntryIo::Data(mut d) => {
                         let expected = d.limit();
-                        if io::copy(&mut d, &mut f)? != expected {
+                        if io::copy(&mut d, &mut f).await? != expected {
                             return Err(other("failed to write entire file"));
                         }
                     }
                     EntryIo::Pad(d) => {
                         // TODO: checked cast to i64
                         let to = SeekFrom::Current(d.limit() as i64);
-                        let size = f.seek(to)?;
-                        f.set_len(size)?;
+                        let size = f.seek(to).await?;
+                        f.set_len(size).await?;
                     }
                 }
             }
-            Ok(f)
-        })()
-        .map_err(|e| {
-            let header = self.header.path_bytes();
-            TarError::new(
-                &format!(
-                    "failed to unpack `{}` into `{}`",
-                    String::from_utf8_lossy(&header),
-                    dst.display()
-                ),
-                e,
-            )
-        })?;
+            Ok::<fs::File, io::Error>(f)
+        }
+            .await
+            .map_err(|e| {
+                let header = self.header.path_bytes();
+                TarError::new(
+                    &format!(
+                        "failed to unpack `{}` into `{}`",
+                        String::from_utf8_lossy(&header),
+                        dst.display()
+                    ),
+                    e,
+                )
+            })?;
 
         if self.preserve_mtime {
             if let Ok(mtime) = self.header.mtime() {
                 let mtime = FileTime::from_unix_time(mtime as i64, 0);
-                filetime::set_file_handle_times(&f, Some(mtime), Some(mtime)).map_err(|e| {
+                filetime::set_file_times(&dst, mtime, mtime).map_err(|e| {
                     TarError::new(&format!("failed to set mtime for `{}`", dst.display()), e)
                 })?;
             }
         }
         if let Ok(mode) = self.header.mode() {
-            set_perms(dst, Some(&mut f), mode, self.preserve_permissions)?;
+            set_perms(dst, Some(&mut f), mode, self.preserve_permissions).await?;
         }
         if self.unpack_xattrs {
-            set_xattrs(self, dst)?;
+            set_xattrs(self, dst).await?;
         }
         return Ok(Unpacked::File(f));
 
-        fn set_perms(
+        async fn set_perms(
             dst: &Path,
-            f: Option<&mut std::fs::File>,
+            f: Option<&mut fs::File>,
             mode: u32,
             preserve: bool,
         ) -> Result<(), TarError> {
-            _set_perms(dst, f, mode, preserve).map_err(|e| {
+            _set_perms(dst, f, mode, preserve).await.map_err(|e| {
                 TarError::new(
                     &format!(
                         "failed to set permissions to {:o} \
@@ -624,9 +666,9 @@ impl<'a> EntryFields<'a> {
         }
 
         #[cfg(any(unix, target_os = "redox"))]
-        fn _set_perms(
+        async fn _set_perms(
             dst: &Path,
-            f: Option<&mut std::fs::File>,
+            f: Option<&mut fs::File>,
             mode: u32,
             preserve: bool,
         ) -> io::Result<()> {
@@ -635,8 +677,8 @@ impl<'a> EntryFields<'a> {
             let mode = if preserve { mode } else { mode & 0o777 };
             let perm = fs::Permissions::from_mode(mode as _);
             match f {
-                Some(f) => f.set_permissions(perm),
-                None => fs::set_permissions(dst, perm),
+                Some(f) => f.set_permissions(perm).await,
+                None => fs::set_permissions(dst, perm).await,
             }
         }
 
@@ -676,11 +718,14 @@ impl<'a> EntryFields<'a> {
         }
 
         #[cfg(all(unix, feature = "xattr"))]
-        fn set_xattrs(me: &mut EntryFields, dst: &Path) -> io::Result<()> {
+        async fn set_xattrs<R: Read + Unpin>(
+            me: &mut EntryFields<R>,
+            dst: &Path,
+        ) -> io::Result<()> {
             use std::ffi::OsStr;
             use std::os::unix::prelude::*;
 
-            let exts = match me.pax_extensions() {
+            let exts = match me.pax_extensions().await {
                 Ok(Some(e)) => e,
                 _ => return Ok(()),
             };
@@ -728,15 +773,15 @@ impl<'a> EntryFields<'a> {
         }
     }
 
-    fn validate_inside_dst(&self, dst: &Path, file_dst: &Path) -> io::Result<PathBuf> {
+    async fn validate_inside_dst(&self, dst: &Path, file_dst: &Path) -> io::Result<PathBuf> {
         // Abort if target (canonical) parent is outside of `dst`
-        let canon_parent = file_dst.canonicalize().map_err(|err| {
+        let canon_parent = file_dst.canonicalize().await.map_err(|err| {
             Error::new(
                 err.kind(),
                 format!("{} while canonicalizing {}", err, file_dst.display()),
             )
         })?;
-        let canon_target = dst.canonicalize().map_err(|err| {
+        let canon_target = dst.canonicalize().await.map_err(|err| {
             Error::new(
                 err.kind(),
                 format!("{} while canonicalizing {}", err, dst.display()),
@@ -757,25 +802,107 @@ impl<'a> EntryFields<'a> {
     }
 }
 
-impl<'a> Read for EntryFields<'a> {
-    fn read(&mut self, into: &mut [u8]) -> io::Result<usize> {
+impl<R: Read + Unpin> Read for EntryFields<R> {
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        into: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        let mut this = self.project();
         loop {
-            match self.data.get_mut(0).map(|io| io.read(into)) {
-                Some(Ok(0)) => {
-                    self.data.remove(0);
+            if this.read_state.is_none() {
+                if this.data.as_ref().is_empty() {
+                    *this.read_state = None;
+                } else {
+                    *this.read_state = Some(this.data.get_mut().remove(0));
                 }
-                Some(r) => return r,
-                None => return Ok(0),
+            }
+
+            let mut read_state = (&mut *this.read_state).take();
+
+            if let Some(ref mut io) = read_state {
+                match Pin::new(io).poll_read(cx, into) {
+                    Poll::Ready(Ok(0)) => {
+                        return Poll::Ready(Ok(0));
+                    }
+                    Poll::Ready(val) => {
+                        *this.read_state = read_state;
+                        return Poll::Ready(val);
+                    }
+                    Poll::Pending => {
+                        *this.read_state = read_state;
+                        return Poll::Pending;
+                    }
+                }
+            } else {
+                return Poll::Ready(Ok(0));
             }
         }
     }
 }
 
-impl<'a> Read for EntryIo<'a> {
-    fn read(&mut self, into: &mut [u8]) -> io::Result<usize> {
-        match *self {
-            EntryIo::Pad(ref mut io) => io.read(into),
-            EntryIo::Data(ref mut io) => io.read(into),
+impl<R: Read + Unpin> Read for EntryIo<R> {
+    #[project]
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        into: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        #[project]
+        match self.project() {
+            EntryIo::Pad(io) => io.poll_read(cx, into),
+            EntryIo::Data(io) => io.poll_read(cx, into),
         }
     }
+}
+
+struct Guard<'a> {
+    buf: &'a mut Vec<u8>,
+    len: usize,
+}
+
+impl Drop for Guard<'_> {
+    fn drop(&mut self) {
+        unsafe {
+            self.buf.set_len(self.len);
+        }
+    }
+}
+
+fn poll_read_all_internal<R: Read + ?Sized>(
+    mut rd: Pin<&mut R>,
+    cx: &mut Context<'_>,
+    buf: &mut Vec<u8>,
+) -> Poll<io::Result<usize>> {
+    let mut g = Guard {
+        len: buf.len(),
+        buf,
+    };
+    let ret;
+    loop {
+        if g.len == g.buf.len() {
+            unsafe {
+                g.buf.reserve(32);
+                let capacity = g.buf.capacity();
+                g.buf.set_len(capacity);
+
+                let buf = &mut g.buf[g.len..];
+                std::ptr::write_bytes(buf.as_mut_ptr(), 0, buf.len());
+            }
+        }
+
+        match async_std::task::ready!(rd.as_mut().poll_read(cx, &mut g.buf[g.len..])) {
+            Ok(0) => {
+                ret = Poll::Ready(Ok(g.len));
+                break;
+            }
+            Ok(n) => g.len += n,
+            Err(e) => {
+                ret = Poll::Ready(Err(e));
+                break;
+            }
+        }
+    }
+
+    ret
 }
