+++ tests/entry.rs
@@ -1,8 +1,8 @@
 extern crate tar;
 extern crate tempfile;
 
-use std::fs::File;
-use std::io::Read;
+use async_std::fs::File;
+use async_std::prelude::*;
 
 use tempfile::Builder;
 
@@ -15,8 +15,8 @@ macro_rules! t {
     };
 }
 
-#[test]
-fn absolute_symlink() {
+#[async_std::test]
+async fn absolute_symlink() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -25,24 +25,24 @@ fn absolute_symlink() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("/bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    t!(ar.unpack(td.path()));
+    t!(ar.unpack(td.path()).await);
 
     t!(td.path().join("foo").symlink_metadata());
 
     let mut ar = tar::Archive::new(&bytes[..]);
     let mut entries = t!(ar.entries());
-    let entry = t!(entries.next().unwrap());
+    let entry = t!(entries.next().await.unwrap());
     assert_eq!(&*entry.link_name_bytes().unwrap(), b"/bar");
 }
 
-#[test]
-fn absolute_hardlink() {
+#[async_std::test]
+async fn absolute_hardlink() {
     let td = t!(Builder::new().prefix("tar").tempdir());
     let mut ar = tar::Builder::new(Vec::new());
 
@@ -51,7 +51,7 @@ fn absolute_hardlink() {
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
@@ -60,18 +60,18 @@ fn absolute_hardlink() {
     // This absolute path under tempdir will be created at unpack time
     t!(header.set_link_name(td.path().join("foo")));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
-    t!(ar.unpack(td.path()));
+    t!(ar.unpack(td.path()).await);
     t!(td.path().join("foo").metadata());
     t!(td.path().join("bar").metadata());
 }
 
-#[test]
-fn relative_hardlink() {
+#[async_std::test]
+async fn relative_hardlink() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -79,7 +79,7 @@ fn relative_hardlink() {
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
@@ -87,19 +87,19 @@ fn relative_hardlink() {
     t!(header.set_path("bar"));
     t!(header.set_link_name("foo"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    t!(ar.unpack(td.path()));
+    t!(ar.unpack(td.path()).await);
     t!(td.path().join("foo").metadata());
     t!(td.path().join("bar").metadata());
 }
 
-#[test]
-fn absolute_link_deref_error() {
+#[async_std::test]
+async fn absolute_link_deref_error() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -108,26 +108,26 @@ fn absolute_link_deref_error() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("/"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo/bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    assert!(ar.unpack(td.path()).is_err());
+    assert!(ar.unpack(td.path()).await.is_err());
     t!(td.path().join("foo").symlink_metadata());
-    assert!(File::open(td.path().join("foo").join("bar")).is_err());
+    assert!(File::open(td.path().join("foo").join("bar")).await.is_err());
 }
 
-#[test]
-fn relative_link_deref_error() {
+#[async_std::test]
+async fn relative_link_deref_error() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -136,27 +136,27 @@ fn relative_link_deref_error() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("../../../../"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo/bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    assert!(ar.unpack(td.path()).is_err());
+    assert!(ar.unpack(td.path()).await.is_err());
     t!(td.path().join("foo").symlink_metadata());
-    assert!(File::open(td.path().join("foo").join("bar")).is_err());
+    assert!(File::open(td.path().join("foo").join("bar")).await.is_err());
 }
 
-#[test]
+#[async_std::test]
 #[cfg(unix)]
-fn directory_maintains_permissions() {
+async fn directory_maintains_permissions() {
     use ::std::os::unix::fs::PermissionsExt;
 
     let mut ar = tar::Builder::new(Vec::new());
@@ -167,22 +167,22 @@ fn directory_maintains_permissions() {
     t!(header.set_path("foo"));
     header.set_mode(0o777);
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    t!(ar.unpack(td.path()));
-    let f = t!(File::open(td.path().join("foo")));
-    let md = t!(f.metadata());
+    t!(ar.unpack(td.path()).await);
+    let f = t!(File::open(td.path().join("foo")).await);
+    let md = t!(f.metadata().await);
     assert!(md.is_dir());
     assert_eq!(md.permissions().mode(), 0o40777);
 }
 
-#[test]
+#[async_std::test]
 #[cfg(not(windows))] // dangling symlinks have weird permissions
-fn modify_link_just_created() {
+async fn modify_link_just_created() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -191,36 +191,36 @@ fn modify_link_just_created() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("bar/foo"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo/bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    t!(ar.unpack(td.path()));
+    t!(ar.unpack(td.path()).await);
 
-    t!(File::open(td.path().join("bar/foo")));
-    t!(File::open(td.path().join("bar/bar")));
-    t!(File::open(td.path().join("foo/foo")));
-    t!(File::open(td.path().join("foo/bar")));
+    t!(File::open(td.path().join("bar/foo")).await);
+    t!(File::open(td.path().join("bar/bar")).await);
+    t!(File::open(td.path().join("foo/foo")).await);
+    t!(File::open(td.path().join("foo/bar")).await);
 }
 
-#[test]
-fn parent_paths_error() {
+#[async_std::test]
+async fn parent_paths_error() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -229,27 +229,27 @@ fn parent_paths_error() {
     t!(header.set_path("foo"));
     t!(header.set_link_name(".."));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo/bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    assert!(ar.unpack(td.path()).is_err());
+    assert!(ar.unpack(td.path()).await.is_err());
     t!(td.path().join("foo").symlink_metadata());
-    assert!(File::open(td.path().join("foo").join("bar")).is_err());
+    assert!(File::open(td.path().join("foo").join("bar")).await.is_err());
 }
 
-#[test]
+#[async_std::test]
 #[cfg(unix)]
-fn good_parent_paths_ok() {
+async fn good_parent_paths_ok() {
     use std::path::PathBuf;
     let mut ar = tar::Builder::new(Vec::new());
 
@@ -259,27 +259,27 @@ fn good_parent_paths_ok() {
     t!(header.set_path(PathBuf::from("foo").join("bar")));
     t!(header.set_link_name(PathBuf::from("..").join("bar")));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(0);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("bar"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
-    t!(ar.unpack(td.path()));
+    t!(ar.unpack(td.path()).await);
     t!(td.path().join("foo").join("bar").read_link());
     let dst = t!(td.path().join("foo").join("bar").canonicalize());
-    t!(File::open(dst));
+    t!(File::open(dst).await);
 }
 
-#[test]
-fn modify_hard_link_just_created() {
+#[async_std::test]
+async fn modify_hard_link_just_created() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -288,33 +288,33 @@ fn modify_hard_link_just_created() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("../test"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(1);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo"));
     header.set_cksum();
-    t!(ar.append(&header, &b"x"[..]));
+    t!(ar.append(&header, &b"x"[..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
 
     let test = td.path().join("test");
-    t!(File::create(&test));
+    t!(File::create(&test).await);
 
     let dir = td.path().join("dir");
-    assert!(ar.unpack(&dir).is_err());
+    assert!(ar.unpack(&dir).await.is_err());
 
     let mut contents = Vec::new();
-    t!(t!(File::open(&test)).read_to_end(&mut contents));
+    t!(t!(File::open(&test).await).read_to_end(&mut contents).await);
     assert_eq!(contents.len(), 0);
 }
 
-#[test]
-fn modify_symlink_just_created() {
+#[async_std::test]
+async fn modify_symlink_just_created() {
     let mut ar = tar::Builder::new(Vec::new());
 
     let mut header = tar::Header::new_gnu();
@@ -323,27 +323,27 @@ fn modify_symlink_just_created() {
     t!(header.set_path("foo"));
     t!(header.set_link_name("../test"));
     header.set_cksum();
-    t!(ar.append(&header, &[][..]));
+    t!(ar.append(&header, &[][..]).await);
 
     let mut header = tar::Header::new_gnu();
     header.set_size(1);
     header.set_entry_type(tar::EntryType::Regular);
     t!(header.set_path("foo"));
     header.set_cksum();
-    t!(ar.append(&header, &b"x"[..]));
+    t!(ar.append(&header, &b"x"[..]).await);
 
-    let bytes = t!(ar.into_inner());
+    let bytes = t!(ar.into_inner().await);
     let mut ar = tar::Archive::new(&bytes[..]);
 
     let td = t!(Builder::new().prefix("tar").tempdir());
 
     let test = td.path().join("test");
-    t!(File::create(&test));
+    t!(File::create(&test).await);
 
     let dir = td.path().join("dir");
-    t!(ar.unpack(&dir));
+    t!(ar.unpack(&dir).await);
 
     let mut contents = Vec::new();
-    t!(t!(File::open(&test)).read_to_end(&mut contents));
+    t!(t!(File::open(&test).await).read_to_end(&mut contents).await);
     assert_eq!(contents.len(), 0);
 }
